backend/requirements.txt
# Django Core
Django==4.2.11
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
django-redis==5.4.0

# Password Hashing (Argon2 - more secure than bcrypt)
argon2-cffi==23.1.0

# CORS Headers
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# Rate Limiting
django-ratelimit==4.1.0

# API Documentation
drf-spectacular==0.27.0

# Web Server (Production)
gunicorn==21.2.0
whitenoise==6.6.0

# PDF Generation
reportlab==4.0.9
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development Tools (optional - uncomment for dev)
# pytest==7.4.4
# pytest-django==4.7.0
# faker==22.5.1
# ipython==8.20.0
